# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API",
    "description": "API",
    "termsOfService": "https://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "https://www.swagger.io/support",
      "email": "neoarmandosousa1010@hotmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Servidor de desenvolvimento"
    },
    {
      "url": "https://projeto-web-full-stack-sacola-development-back-end.vercel.app/",
      "description": "Servidor de produção"
    }
  ],
  "paths": {
    "/api/v1/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Client_Token",
            "in": "header",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não Autorizado: A autenticação é necessária e falhou ou não foi fornecida.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflito: O usuário já existe",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Cria um usuário",
        "description": "Cria um usuário",
        "operationId": "createUser"
      },
      "get": {
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Usuários encontrados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "users": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não Autorizado: A autenticação é necessária e falhou ou não foi fornecida.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Busca todos os usuários",
        "description": "Busca todos os usuários",
        "operationId": "getUsers"
      }
    },
    "/api/v1/user": {
      "get": {
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário encontrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não Autorizado: A autenticação é necessária e falhou ou não foi fornecida.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Busca um usuário",
        "description": "Busca um usuário",
        "operationId": "getUser"
      },
      "put": {
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Client_Token",
            "in": "header",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não Autorizado: A autenticação é necessária e falhou ou não foi fornecida.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado: O usuário não foi encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Atualiza um usuário",
        "description": "Atualiza um usuário"
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Client_Token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário deletado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não Autorizado: A autenticação é necessária e falhou ou não foi fornecida.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado: O usuário não foi encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Deleta um usuário",
        "description": "Deleta um usuário"
      }
    },
    "/api/v1/user/games": {
      "put": {
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "Client_Token",
            "in": "header",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "game": {
                    "type": "object",
                    "description": "jogo do usuário"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Jogo adicionado ao usuário com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não Autorizado: A autenticação é necessária e falhou ou não foi fornecida.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflito: O jogo já existe no usuário",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de erro"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Adiciona um jogo ao usuário",
        "description": "Adiciona um jogo ao usuário",
        "operationId": "addGameToUser"
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "properties": {
          "_id": {
            "type": "string",
            "description": "Id do usuário"
          },
          "name": {
            "type": "string",
            "description": "Nome do usuário"
          },
          "username": {
            "type": "string",
            "description": "apelido do usuário"
          },
          "email": {
            "type": "string",
            "description": "email do usuário"
          },
          "photo": {
            "type": "string",
            "description": "foto do usuário"
          },
          "userFriends": {
            "type": "object",
            "description": "amigos do usuário"
          },
          "userGames": {
            "type": "object",
            "description": "jogos do usuário"
          },
          "storeKeys": {
            "type": "object",
            "description": "chaves das lojas do usuário"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}